+++ src/codeql.test.ts
@@ -16,6 +16,7 @@ import { CommandInvocationError } from "./cli-errors";
 import * as codeql from "./codeql";
 import { AugmentationProperties, Config } from "./config-utils";
 import * as defaults from "./defaults.json";
+import { DocUrl } from "./doc-url";
 import { Language } from "./languages";
 import { getRunnerLogger } from "./logging";
 import { ToolsSource } from "./setup-codeql";
@@ -906,11 +907,10 @@ test("runTool summarizes autobuilder errors", async (t) => {
   await t.throwsAsync(
     async () => await codeqlObject.runAutobuild(stubConfig, Language.java),
     {
-      instanceOf: CommandInvocationError,
+      instanceOf: util.ConfigurationError,
       message:
         "We were unable to automatically build your code. Please provide manual build steps. " +
-        "For more information, see " +
-        "https://docs.github.com/en/code-security/code-scanning/troubleshooting-code-scanning/automatic-build-failed. " +
+        `See ${DocUrl.AUTOMATIC_BUILD_FAILED} for more information. ` +
         "Encountered the following error: Start of the error message\n" +
         "  Some more context about the error message\n" +
         "  continued\n" +
@@ -934,11 +934,10 @@ test("runTool truncates long autobuilder errors", async (t) => {
   await t.throwsAsync(
     async () => await codeqlObject.runAutobuild(stubConfig, Language.java),
     {
-      instanceOf: CommandInvocationError,
+      instanceOf: util.ConfigurationError,
       message:
         "We were unable to automatically build your code. Please provide manual build steps. " +
-        "For more information, see " +
-        "https://docs.github.com/en/code-security/code-scanning/troubleshooting-code-scanning/automatic-build-failed. " +
+        `See ${DocUrl.AUTOMATIC_BUILD_FAILED} for more information. ` +
         "Encountered the following error: " +
         `${Array.from({ length: 10 }, (_, i) => `line${i + 1}`).join(
           "\n",